Prevent search path from leaking outside of tests
Before using mdfind we were only detecting Xcodes in /Applications.
Since we're 1) using Spotlight and 2) only detecting when sourced it
should be fine to search everywhere Spotlight indexes.

The test suite however still needs a way to override the search path so
it doesn't detect real Xcodes installed on a system. To make this work
we define `xcode_search_path` which the test suite overrides. `chxcode`
immediately unsets this variable when it's done with it. Preventing it
from leaking into a user's shell.